grunge_livelike /requirements.txt
appdirs==1.4.4
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
flake8==7.1.1
furl==2.1.3
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
orderedmultidict==1.0.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2024.2
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
toml==0.10.2
tomli==2.0.2
typing_extensions==4.12.2
